requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.9
django-session-timeout==0.1.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-decouple==3.8
requests==2.32.3
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
sslcommerz-lib==1.0
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0